backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn
python-multipart==0.0.9
PyJWT[crypto]==2.9.0
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
matplotlib==3.9.2
numpy
aiohttp>=3.9.0
pypdf==5.0.1
python-docx==0.8.11
